+++ doc/changelog.rst
@@ -72,6 +72,8 @@ Xtheme
 Classic Gray Theme
 ~~~~~~~~~~~~~~~~~~
 
+- Remove ``ProductCrossSellType.COMPUTED`` from cross-sells plugin
+- Update cross-sells plugin to use ``ProductCrossSellType.BOUGHT_WITH``
 - Render prices with the new price rendering template tags
 - Show error messages while adding products to basket
 - Add Coupon use possibility to basket page